backend/requirements.txt
celery==5.0.3
celery-progress==0.0.14
dataclasses==0.6
dj-database-url==0.5.0
Django==3.2.10
django-jazzmin==2.4.8
gunicorn==20.0.4
python-arango==6.1.0
python-dateutil==2.8.1
python-decouple==3.3
python-magic==0.4.18
redis==3.5.3
django-redis==5.2.0
regex==2020.11.13
spacy==2.3.4
textract==1.6.4
Unipath==1.1
whitenoise==5.2.0
psycopg2-binary==2.8.5
django-two-factor-auth==1.13.2
phonenumbers==8.12.21
YubiOTP==1.0.0 
django-otp-yubikey==1.0.0.post1
django-crispy-forms==1.13.0
coverage==5.5
django-haystack==3.0
Whoosh==2.7.4
django-user-visit==0.5.1
django-tenants==3.4.1
django-prometheus==2.1.0
prometheus-client==0.11.0
tenant-schemas-celery==1.0.1
flake8==3.9.2
selenium==3.141.0
typing==3.7.4.3
passwordgenerator==1.4
django-honeypot==1.0.2
typing-extensions==4.0.1
django-countries==7.2.1
django-formtools==2.3
crispy-bootstrap5==0.6
django-bootstrap-v5==1.0.7
django-ses==2.3.1
django-limits==0.0.6
stripe==2.64.0
dj-stripe==2.5.1
django-defender==0.9.1
django-csp==3.7
django-ratelimit==3.0.1
django-admin-honeypot==1.1.0
django-utils-six==2.0
django-pgcrypto-fields==2.6.0
channels==3.0.5
channels-redis==3.3.1










